+++ packages/claims-controller/CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Added `clearClaimsState` to clear/reset the claims controller state. ([#7780](https://github.com/MetaMask/core/pull/7780))
+- Added new public method, `clearState` to clear/reset the claims controller state. ([#7780](https://github.com/MetaMask/core/pull/7780))
 
 ### Changed